requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0.1
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
django-dbbackup==3.3.0
django-anymail[mailgun]==8.5
psycopg2-binary==2.9.3
PyJWT==2.3.0
pytz==2021.3
requests==2.27.1
twilio==7.5.0
django-phonenumber-field==6.0.0
phonenumbers==8.12.41
pillow==9.0.0
django-storages==1.12.3
django-sendsms==0.5
boto3==1.20.41
awscli==1.22.41
